Relocate the egress-decision core into mvm-contract
`policy/projection.rs` decides every workload's egress: `wasi_allows`
projects a `NetworkPolicy` onto a WASI grant set, and
`CanonicalEgress::permits` is the counterpart the L3/L4 enforcers agree
with. It was pure decision logic with no I/O and no resolver, sitting
in `mvm-core` only because Increment 3 was scoped to DTOs.

Move it down verbatim so a no_std/edge/browser consumer runs the host's
gate rather than a second copy of it. The file is a `git mv` plus nine
lines: `std::net` -> `core::net`, the `alloc` prelude, and
`std::str::FromStr` -> `core::str::FromStr`. `mvm-core` re-exports the
module at `policy::projection`, so all ~20 call sites -- `EgressGate`,
mvm-net's L3 admit, mvm-hostd's proxy and DNS handler, mvm-conformance
-- are unchanged, and `wasm_egress_witness.rs` stays green unmodified.

`is_mandatory_deny`, `mandatory_deny_ranges` and `unmap_v4_mapped`
moved with it, for the same reason and by the same re-export: they are
pure predicates over the already-relocated `MANDATORY_DENY_RANGES`
const, and projection decides with them. The iptables script
generators, which are genuinely host-only, stayed in `mvm-core`.

ipnet joins mvm-contract with `default-features = false`; without its
`std` feature its `AddrParseError` implements `core::error::Error`,
which is what `ProjectionError`'s thiserror derive needs to carry it as
a source.

Side effect: the 54 projection tests and 10 mandatory-deny tests now
run under `wasm32-wasip1` (651 -> 715) and the decision core builds for
`riscv32imac-unknown-none-elf`. Neither had wasm coverage before.

No serde shape changes; the projection types carry no serde at all. No
claim-catalog witness -- this touches no ADR-001 row.
